+++ src/OpenTelemetry.Api/Internal/SpanHelper.cs
@@ -23,7 +23,7 @@ namespace OpenTelemetry.Trace
     {
         /// <summary>
         /// Helper method that populates span properties from http status code according
-        /// to https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/trace/semantic_conventions/http.md#status.
+        /// to https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/http.md#status.
         /// </summary>
         /// <param name="httpStatusCode">Http status code.</param>
         /// <returns>Resolved span <see cref="Status"/> for the Http status code.</returns>