+++ src/OpenTelemetry/Resources/ResourceBuilderExtensions.cs
@@ -42,7 +42,7 @@ namespace OpenTelemetry.Resources
         /// <summary>
         /// Adds service information to a <see cref="ResourceBuilder"/>
         /// following <a
-        /// href="https://github.com/open-telemetry/opentelemetry-specification/tree/master/specification/resource/semantic_conventions#service">semantic
+        /// href="https://github.com/open-telemetry/opentelemetry-specification/tree/main/specification/resource/semantic_conventions#service">semantic
         /// conventions</a>.
         /// </summary>
         /// <param name="resourceBuilder"><see cref="ResourceBuilder"/>.</param>
@@ -132,7 +132,7 @@ namespace OpenTelemetry.Resources
         /// <summary>
         /// Adds resource attributes parsed from an environment variable to a
         /// <see cref="ResourceBuilder"/> following the <a
-        /// href="https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/resource/sdk.md#specifying-resource-information-via-an-environment-variable">Resource
+        /// href="https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/resource/sdk.md#specifying-resource-information-via-an-environment-variable">Resource
         /// SDK</a>.
         /// </summary>
         /// <param name="resourceBuilder"><see cref="ResourceBuilder"/>.</param>