+++ src/OpenTelemetry.Exporter.Jaeger/CHANGELOG.md
@@ -5,8 +5,8 @@
 * Changed `JaegerExporter` class and constructor from internal to public.
   ([#1612](https://github.com/open-telemetry/opentelemetry-dotnet/issues/1612))
 
-* In `JaegerExporterOptions`: Exporter options now include a switch for
-  Batch vs Simple exporter, and settings for batch exporting properties.
+* In `JaegerExporterOptions`: Exporter options now include a switch for Batch vs
+  Simple exporter, and settings for batch exporting properties.
 
 * Jaeger will now set the `error` tag when `otel.status_code` is set to `ERROR`.
   ([#1579](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1579) &
@@ -19,6 +19,10 @@
 * Span Event.Name will now be populated as the `event` field on Jaeger Logs
   instead of `message`.
   ([#1609](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1609))
+
+* `JaegerExporter` batch format has changed to be compliant with the spec. This
+  may impact the way spans are displayed in Jaeger UI.
+  ([#1732](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1732))
 
 ## 1.0.0-rc1.1